Split guardrails into focused package
