+++ src/omnicoreagent/core/guardrails/__init__.py
@@ -1,0 +1,23 @@
+from omnicoreagent.core.guardrails.engine import DetectionEngine
+from omnicoreagent.core.guardrails.guard import (
+    PromptInjectionGuard,
+    create_guard,
+    quick_check,
+)
+from omnicoreagent.core.guardrails.models import (
+    DetectionConfig,
+    DetectionResult,
+    ThreatLevel,
+)
+from omnicoreagent.core.guardrails.patterns import PatternManager
+
+__all__ = [
+    "DetectionConfig",
+    "DetectionEngine",
+    "DetectionResult",
+    "PatternManager",
+    "PromptInjectionGuard",
+    "ThreatLevel",
+    "create_guard",
+    "quick_check",
+]
